# api/openapi.json
{
  "components": {
    "schemas": {
...
      "BriefV1": {
        "additionalProperties": false,
        "properties": {
          "act_candidates": {
            "items": {
              "$ref": "#/components/schemas/TapeRowV1"
            },
            "title": "Act Candidates",
            "type": "array"
          },
          "funnel": {
            "$ref": "#/components/schemas/FunnelV1"
          },
          "leading_groups": {
            "items": {
              "$ref": "#/components/schemas/GroupSummaryV1"
            },
            "title": "Leading Groups",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ViewMetaV1"
          },
          "portfolio_heat_status": {
            "const": "Unavailable \u2014 portfolio context not implemented",
            "default": "Unavailable \u2014 portfolio context not implemented",
            "title": "Portfolio Heat Status",
            "type": "string"
          },
          "regime_reason": {
            "title": "Regime Reason",
            "type": "string"
          },
          "regime_state": {
            "title": "Regime State",
            "type": "string"
          },
          "sleeves": {
            "items": {
              "$ref": "#/components/schemas/SleeveViewV1"
            },
            "title": "Sleeves",
            "type": "array"
          },
          "weakening_groups": {
            "items": {
              "$ref": "#/components/schemas/GroupSummaryV1"
            },
            "title": "Weakening Groups",
            "type": "array"
          }
        },
        "required": [
          "meta",
          "regime_state",
          "regime_reason",
          "sleeves",
          "funnel",
          "leading_groups",
          "weakening_groups",
          "act_candidates"
        ],
        "title": "BriefV1",
        "type": "object"
      },
...
      "DecisionState": {
        "enum": [
          "NONE",
          "WATCH",
          "TRADE",
          "ACT"
        ],
        "title": "DecisionState",
        "type": "string"
      },
...
      "Direction": {
        "enum": [
          "LONG",
          "SHORT"
        ],
        "title": "Direction",
        "type": "string"
      },
...
      "ErrorV1": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ErrorFieldV1"
            },
            "title": "Fields",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "mode": {
            "enum": [
              "FIXTURE",
              "LOCAL_SNAPSHOT"
            ],
            "title": "Mode",
            "type": "string"
          },
          "mode_label": {
            "enum": [
              "SYNTHETIC FIXTURE",
              "LOCAL SNAPSHOT"
            ],
            "title": "Mode Label",
            "type": "string"
          },
          "schema_version": {
            "const": "workstation-error-v1",
            "default": "workstation-error-v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "mode_label",
          "code",
          "message"
        ],
        "title": "ErrorV1",
        "type": "object"
      },
...
      "FunnelV1": {
        "additionalProperties": false,
        "properties": {
          "ACT": {
            "minimum": 0.0,
            "title": "Act",
            "type": "integer"
          },
          "NONE": {
            "minimum": 0.0,
            "title": "None",
            "type": "integer"
          },
          "TRADE": {
            "minimum": 0.0,
            "title": "Trade",
            "type": "integer"
          },
          "WATCH": {
            "minimum": 0.0,
            "title": "Watch",
            "type": "integer"
          }
        },
        "required": [
          "NONE",
          "WATCH",
          "TRADE",
          "ACT"
        ],
        "title": "FunnelV1",
        "type": "object"
      },
...
      "RulesViewV1": {
        "additionalProperties": false,
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ViewMetaV1"
          },
          "rules_fingerprint": {
            "title": "Rules Fingerprint",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/RuleSectionV1"
            },
            "title": "Sections",
            "type": "array"
          },
          "status": {
            "const": "experimental_uncalibrated",
            "default": "experimental_uncalibrated",
            "title": "Status",
            "type": "string"
          },
          "versions": {
            "$ref": "#/components/schemas/VersionsV1"
          }
        },
        "required": [
          "meta",
          "versions",
          "rules_fingerprint",
          "sections"
        ],
        "title": "RulesViewV1",
        "type": "object"
      },
...
      "SizerRequestV1": {
        "additionalProperties": false,
        "properties": {
          "account_equity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Equity"
          },
          "available_buying_power": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Buying Power"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "entry": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry"
          },
          "stop": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "direction",
          "account_equity",
          "available_buying_power",
          "entry",
          "stop"
        ],
        "title": "SizerRequestV1",
        "type": "object"
      },
      "SizerResponseV1": {
        "additionalProperties": false,
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ViewMetaV1"
          },
          "result": {
            "$ref": "#/components/schemas/SizingResultV1"
          }
        },
        "required": [
          "meta",
          "result"
        ],
        "title": "SizerResponseV1",
        "type": "object"
      },
...
      "SleeveViewV1": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "name",
          "state",
          "score",
          "reasons"
        ],
        "title": "SleeveViewV1",
        "type": "object"
      },
...
      "SymbolDetailV1": {
        "additionalProperties": false,
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ViewMetaV1"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/SymbolRecordV1"
            },
            "title": "Records",
            "type": "array"
          }
        },
        "required": [
          "meta",
          "records"
        ],
        "title": "SymbolDetailV1",
        "type": "object"
      },
...
      "TapeRowV1": {
        "additionalProperties": false,
        "properties": {
          "RS_comp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rs Comp"
          },
          "RS_rotation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rs Rotation"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionState"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "earnings": {
            "title": "Earnings",
            "type": "string"
          },
          "extension_atr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Atr"
          },
          "group_rank": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Rank"
          },
          "has_veto": {
            "title": "Has Veto",
            "type": "boolean"
          },
          "percentile_21": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentile 21"
          },
          "percentile_5": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentile 5"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonV1"
            },
            "title": "Reasons",
            "type": "array"
          },
          "return_21": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return 21"
          },
          "return_5": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return 5"
          },
          "rotation_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rotation Delta"
          },
          "setups": {
            "items": {
              "$ref": "#/components/schemas/SetupSummaryV1"
            },
            "title": "Setups",
            "type": "array"
          },
          "structure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structure"
          },
          "sub_industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Industry"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "direction",
          "display_name",
          "price",
          "structure",
          "RS_comp",
          "return_5",
          "percentile_5",
          "return_21",
          "percentile_21",
          "RS_rotation",
          "rotation_delta",
          "sub_industry",
          "group_rank",
          "setups",
          "extension_atr",
          "decision",
          "earnings",
          "reasons",
          "has_veto"
        ],
        "title": "TapeRowV1",
        "type": "object"
      },
      "TapeV1": {
        "additionalProperties": false,
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ViewMetaV1"
          },
          "order": {
            "enum": [
              "asc",
              "desc"
            ],
            "title": "Order",
            "type": "string"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/TapeRowV1"
            },
            "title": "Rows",
            "type": "array"
          },
          "sort": {
            "title": "Sort",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "meta",
          "rows",
          "total",
          "page",
          "page_size",
          "pages",
          "sort",
          "order"
        ],
        "title": "TapeV1",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Aperture Workstation",
    "version": "workstation-api-v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/brief": {
      "get": {
        "operationId": "brief_api_v1_brief_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Brief"
      }
    },
...
    "/api/v1/rules": {
      "get": {
        "operationId": "rules_api_v1_rules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulesViewV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Rules"
      }
    },
    "/api/v1/sizer": {
      "post": {
        "operationId": "sizer_api_v1_sizer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SizerRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SizerResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Sizer"
      }
    },
    "/api/v1/symbols/{symbol}": {
      "get": {
        "operationId": "symbol_detail_api_v1_symbols__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SymbolDetailV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Symbol Detail"
      }
    },
    "/api/v1/tape": {
      "get": {
        "operationId": "tape_api_v1_tape_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 100000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "symbol",
              "enum": [
                "symbol",
                "price",
                "RS_comp",
                "RS_rotation",
                "rotation_delta",
                "extension_atr",
                "decision",
                "group_rank"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "NONE",
                    "WATCH",
                    "TRADE",
                    "ACT"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "in": "query",
            "name": "structure",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "NEUTRAL",
                    "EMERGING",
                    "UPTREND",
                    "DETERIORATING",
                    "DECLINE"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Structure"
            }
          },
          {
            "in": "query",
            "name": "setup",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "EP",
                    "CONTRACTION",
                    "TREND_PULLBACK",
                    "RANGE"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Setup"
            }
          },
          {
            "in": "query",
            "name": "min_rs_comp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Rs Comp"
            }
          },
          {
            "in": "query",
            "name": "min_rs_rotation",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Rs Rotation"
            }
          },
          {
            "in": "query",
            "name": "group",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 160,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group"
            }
          },
          {
            "in": "query",
            "name": "veto",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Veto"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TapeV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorV1"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Tape"
      }
    }
  }
}